Seperated Header Hero and SM icons
