+++ src/components/MagazineScroller.css
@@ -77,7 +77,7 @@
 .circle-cards-container ol li::after {
   content: "";
   filter: drop-shadow(-0.25rem 0.25rem 0.0675rem rgba(0, 0, 0, 0.75)) blur(5px);
-  
+
 }
 
 .circle-cards-container .circle-img {
@@ -94,7 +94,7 @@
   height: 100%;
   object-fit: cover;
   border-radius: 50%; /* Ensure image follows circular shape */
-  
+
 }
 
 .circle-cards-container .circle-cards-heading {
@@ -139,7 +139,7 @@
   backdrop-filter: blur(10px);
    display: inline-flex; /* keeps text + ::after inline */
   align-items: center;  /* vertically centers them */
-  gap: 0.5rem;  
+  gap: 0.5rem;
 }
 
 .trending-box::before {
@@ -177,7 +177,7 @@
 
 .trending-label {
   opacity: 0.9;
-  
+
 }
 
 .typewriter-text {
@@ -246,13 +246,13 @@
 }
 @keyframes pulseGlow {
   0% {
-    box-shadow: 
+    box-shadow:
       0 0 6px rgba(0, 174, 255, 0.6),
       0 0 12px rgba(0, 174, 255, 0.4),
       0 0 18px rgba(0, 174, 255, 0.2);
   }
   100% {
-    box-shadow: 
+    box-shadow:
       0 0 12px rgba(0, 174, 255, 0.8),
       0 0 20px rgba(0, 174, 255, 0.5),
       0 0 28px rgba(0, 174, 255, 0.3);
