+++ src/components/SocialMedia.css
@@ -1,0 +1,84 @@
+/* Social Media Sidebar */
+.acm-social-sidebar {
+    position: fixed;
+    left: 30px;
+    top: 50%;
+    transform: translateY(-50%) translateX(-100px);
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    z-index: 1000;
+    opacity: 0;
+    transition: all 0.8s cubic-bezier(0.4, 0, 0.2, 1);
+}
+
+.acm-social-sidebar.animate {
+    transform: translateY(-50%) translateX(0);
+    opacity: 1;
+    animation: acm-sidebarSlide 0.8s cubic-bezier(0.4, 0, 0.2, 1) 1s forwards;
+}
+
+@keyframes acm-sidebarSlide {
+    0% {
+        transform: translateY(-50%) translateX(-100px);
+        opacity: 0;
+    }
+    100% {
+        transform: translateY(-50%) translateX(0);
+        opacity: 1;
+    }
+}
+
+.acm-social-icon {
+    width: 50px;
+    height: 50px;
+    border-radius: 50%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    color: white;
+    text-decoration: none;
+    font-size: 20px;
+    transition: all 0.3s ease;
+    backdrop-filter: blur(10px);
+    border: 1px solid rgba(255, 255, 255, 0.2);
+}
+
+.acm-social-icon:hover {
+    transform: scale(1.1) translateY(-5px);
+    box-shadow: 0 10px 25px rgba(0, 0, 0, 0.3);
+}
+
+.acm-facebook {
+    background: linear-gradient(135deg, #1877f2, #42a5f5);
+}
+.acm-instagram {
+    background: linear-gradient(135deg, #e4405f, #f77737, #fccc63);
+}
+.acm-linkedin {
+    background: linear-gradient(135deg, #0077b5, #00a0dc);
+}
+.acm-twitter {
+    background: linear-gradient(135deg, #1da1f2, #0d8bd9);
+}
+
+/* Responsive Design */
+@media (max-width: 768px) {
+    .acm-social-sidebar {
+        left: 20px;
+        gap: 15px;
+    }
+
+    .acm-social-icon {
+        width: 40px;
+        height: 40px;
+        font-size: 16px;
+    }
+}
+
+@media (max-width: 480px) {
+    .acm-social-sidebar {
+        left: 15px;
+        gap: 12px;
+    }
+}
